# backend/cmd/app/docs/swagger.yaml
basePath: /
definitions:
...
  normaladmin_backend_internal_models.Member:
    properties:
      avatar:
        type: string
      birthday:
        type: string
      created_at:
        type: string
      email:
        type: string
      gender:
        description: 0-未知 1-男 2-女
        type: integer
      id:
        type: integer
      last_login_ip:
        type: string
      last_login_time:
        type: string
      level_id:
        type: integer
      mobile:
        type: string
      nickname:
        type: string
      points:
        type: integer
      status:
        description: 0-禁用 1-启用 2-黑名单
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  normaladmin_backend_pkg_utils_response.ResponseData:
    properties:
      data: {}
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a backend server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend API
  version: "1.0"
paths:
...
  /gam/members:
    get:
      consumes:
      - application/json
      description: 获取会员列表,支持分页和筛选
      parameters:
      - description: 用户名
        in: query
        name: username
        type: string
      - description: 手机号
        in: query
        name: mobile
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 排序字段，多个字段用逗号分隔
        example: '"id,created_at"'
        in: query
        name: sortField
        type: string
      - description: 排序方式(asc/desc)，多个排序方式用逗号分隔
        example: '"desc,asc"'
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    members:
                      items:
                        $ref: '#/definitions/normaladmin_backend_internal_models.Member'
                      type: array
                    total:
                      type: integer
                  type: object
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 获取会员列表
      tags:
      - 会员管理
    post:
      consumes:
      - application/json
      description: 创建新的会员账号
      parameters:
      - description: 会员信息
        in: body
        name: member
        required: true
        schema:
          $ref: '#/definitions/normaladmin_backend_internal_models.Member'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    member:
                      $ref: '#/definitions/normaladmin_backend_internal_models.Member'
                  type: object
              type: object
        "400":
          description: 无效的请求参数
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 创建会员
      tags:
      - 会员管理
  /gam/members/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除会员
      parameters:
      - description: 会员ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: 无效的会员ID
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "404":
          description: 会员不存在
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 删除会员
      tags:
      - 会员管理
    put:
      consumes:
      - application/json
      description: 更新会员信息
      parameters:
      - description: 会员ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 会员信息
        in: body
        name: member
        required: true
        schema:
          $ref: '#/definitions/normaladmin_backend_internal_models.Member'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    member:
                      $ref: '#/definitions/normaladmin_backend_internal_models.Member'
                  type: object
              type: object
        "400":
          description: 无效的请求参数
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "404":
          description: 会员不存在
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 更新会员
      tags:
      - 会员管理
  /gam/members/{id}/status:
    put:
      consumes:
      - application/json
      description: 更新会员的启用/禁用状态（0：禁用，1：启用）
      parameters:
      - description: 会员ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 状态信息
        in: body
        name: status
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: 无效的请求参数
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "404":
          description: 会员不存在
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 更新会员状态
      tags:
      - 会员管理
  /gam/members/check-field:
    get:
      consumes:
      - application/json
      description: 检查会员的某个字段值是否唯一（如用户名、手机号等）
      parameters:
      - description: 字段名
        example: '"username"'
        in: query
        name: field
        required: true
        type: string
      - description: 字段值
        example: '"test"'
        in: query
        name: value
        required: true
        type: string
      - description: 排除的ID
        in: query
        minimum: 1
        name: excludeId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    unique:
                      type: boolean
                  type: object
              type: object
        "400":
          description: 无效的请求参数
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 检查会员字段唯一性
      tags:
      - 会员管理
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"